// react/src/gen/environment_pb.d.ts
export declare enum EnvironmentType {
  /**
   * @generated from enum value: LOCAL = 0;
   */
  LOCAL = 0,

  /**
   * @generated from enum value: DOCKER = 1;
   */
  DOCKER = 1,

  /**
   * @generated from enum value: SSH = 2;
   */
  SSH = 2,
}